Move stuff around in importer
